refactor: reduce dependencies of `query` crate
While we need the full `datafusion` department for testing, we only need
`datafusion-execution` for the actual compilation of the `query` crate.
This somewhat speeds up the compilation of downstream consumers like
InfluxDB since they don't need to wait for `datafusion` to be compiled
before they can compile `query` (that's automatically handled by cargo).